updating dashboard visibility status on the same page and removing routing
