+++ src/app/page.tsx
@@ -2,25 +2,27 @@
 "use client";
 
 import { useState } from "react";
-import { useRouter } from "next/navigation";
 import { Button } from "@/components/ui/button"; // Import your UI components
 import { Input } from "@/components/ui/input";
+import AdminDashboard from "@/components/admin/dashboard";
 
 export default function HomePage() {
   const [code, setCode] = useState("");
-  const router = useRouter();
+  const [showDashboard, setShowDashboard] = useState(false);
 
   const handleSubmit = (e: React.FormEvent) => {
     e.preventDefault();
     // Check if the code matches your criteria
     if (code === "MTx20") {
-      router.push("/admin/dashboard"); // Redirect to dashboard
+      setShowDashboard(true); // Show dashboard
     } else {
       alert("Invalid code. Please try again.");
     }
   };
 
-  return (
+  return showDashboard ? (
+    <AdminDashboard />
+  ) : (
     <div className="flex flex-col items-center justify-center min-h-screen">
       <h1 className="text-3xl font-bold mb-4">Welcome to Our Website</h1>
       <form onSubmit={handleSubmit} className="space-y-4">
